Letter Count V4

+++ A1/LetterCount.java
@@ -25,9 +25,9 @@ public class WordCount {
     }
 }	
     public static class Reduce extends MapReduceBase implements Reducer<Text, IntWritable, Text, IntWritable> {
-        private LongWritable result = new LongWritable();
+        private IntWritable result = new IntWritable();
 
-        public void reduce(Text key, Iterator<IntWritable> values, OutputCollector<Text, LongWritable> output, Reporter reporter) throws IOException {
+        public void reduce(Text key, Iterator<IntWritable> values, OutputCollector<Text, IntWritable> output, Reporter reporter) throws IOException {
             int sum = 0;
             while (values.hasNext()) {
                 sum += values.next().get();
